[CBR-344] Create a cardano-sl-demo package from 'test/Integration/Clients.hs'
This enables us to run a demo cluster at will, without any shell script or nix setup. It also factor out a bit some code from the integration tests to
keep them focused on .. integration tests
