+++ demo/cardano-sl-demo.cabal
@@ -1,0 +1,71 @@
+name:                cardano-sl-demo
+version:             1.0.0
+synopsis:            Tooling to easily run a local demo cluster
+description:         Please see README.md
+homepage:            https://github.com/input-output-hk/cardano-sl/tree/demo/README.md
+author:              IOHK Engineering Team
+maintainer:          operations@iohk.io
+copyright:           2018 IOHK
+license:             MIT
+license-file:        LICENSE
+category:            Web
+build-type:          Simple
+extra-source-files:  README.md
+cabal-version:       >=1.10
+
+library
+  default-language:    Haskell2010
+  ghc-options:         -Wall
+
+  hs-source-dirs:      src
+  exposed-modules:     Cardano.Wallet.Demo
+
+  build-depends:      base
+                    , async
+                    , bytestring
+                    , cardano-sl
+                    , cardano-sl-core
+                    , cardano-sl-crypto
+                    , cardano-sl-infra
+                    , cardano-sl-util
+                    , cardano-sl-wallet
+                    , cardano-sl-wallet-new
+                    , containers
+                    , http-client
+                    , http-types
+                    , lens
+                    , log-warper
+                    , mtl
+                    , optparse-applicative
+                    , random
+                    , stm
+                    , time
+                    , time-units
+                    , universum
+                    , x509-store
+                    , parsec
+                    , attoparsec
+
+  default-extensions: NoImplicitPrelude
+                      FlexibleContexts
+                      OverloadedStrings
+                      RecordWildCards
+                      ScopedTypeVariables
+
+test-suite demo
+  default-language:   Haskell2010
+  ghc-options:        -Wall
+  type:               exitcode-stdio-1.0
+  hs-source-dirs:     test
+  main-is:            Main.hs
+
+  default-extensions: NoImplicitPrelude
+                      FlexibleContexts
+                      OverloadedStrings
+                      RecordWildCards
+                      ScopedTypeVariables
+
+  build-depends:      base
+                    , async
+                    , cardano-sl-demo
+                    , universum
