+++ demo/state-demo/tls/server.crt
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDjjCCAnagAwIBAgIBAjANBgkqhkiG9w0BAQsFADBDMRgwFgYDVQQKDA9JbnB1
+dCBPdXRwdXQgSEsxJzAlBgNVBAMMHkNhcmRhbm8gU0wgU2VsZi1TaWduZWQgUm9v
+dCBDQTAeFw0xODA4MDExNjE2MDVaFw0xOTA4MDExNjE3MDVaMDsxGDAWBgNVBAoM
+D0lucHV0IE91dHB1dCBISzEfMB0GA1UEAwwWQ2FyZGFubyBTTCBTZXJ2ZXIgTm9k
+ZTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAMFmUscVUwMfQxBJE7my
+wKOrEYBaFylqu7C0aG76wAQA21//8g/UBzBqyKWG7YNneNuUjaqTk2WlK/ojUu/e
+Tb1QT+8E7saRAF/chRap/keXOVIVaQTJ96CR2M1EsdiXOMUpLo12/tbUvsSNIB9V
+hhKAWl+fn7fNPvjP2QAd/YfCB3O7WgcuETtAoCOInUr5+D9MoSK+FBVyE0H2aRwt
+Wk3meFaA8oVgzwVSymiAb7hSYmk6+/Pe7qaZL6+I5AIWtliFbcP1+2s19XFfFqhM
+q0y0R424BhBLNCftaELekCtueYfHuKucL2cunfDQMheCdNnP5PLNLWQlzH1II6dW
+1+MCAwEAAaOBlDCBkTA7BgNVHREENDAygglsb2NhbGhvc3SCFWxvY2FsaG9zdC5s
+b2NhbGRvbWFpboIJMTI3LjAuMC4xggM6OjEwDwYDVR0PAQH/BAUDAwegADATBgNV
+HSUEDDAKBggrBgEFBQcDATAMBgNVHRMEBTADAQEAMA0GA1UdDgQGBASEXf7CMA8G
+A1UdIwQIMAaABGwLqOgwDQYJKoZIhvcNAQELBQADggEBAJMiCjQ//QAdhBKasmRN
+6Ucexv6eBbqqGooRoncdVnCyZ3marwYAGyubZ9oagm/v/+KRlblFCBaHuGpSiY8h
+6zyWF6h0GZDSL7J44GZGHvhzz06srt9qcId1hW/7xKTI6mOnaE6sTCp8WRL/cDJo
+WMe9YMyvtuHNLBMypv4NUOSCt7y5v1Ft47+CVZB3CO73K4CjRkiq5AWEIBKcqN27
+TZL6zFGOwa87W4zs2h3TD9EG/jDuHpQvoG0JBNzIx7rRDajDdZVfPrxNfZQTIn32
+ZjRGK6n86b0RoErgZeU+SB4PaLYqes6aCcc2lbGfm5i4PoqdIvvLHIbVcaGofa9x
+zu8=
+-----END CERTIFICATE-----
