+++ wallet-new/cardano-sl-wallet-new.cabal
@@ -245,6 +245,7 @@ executable cardano-node
                        -Wall
                        -O2
   build-depends:       base
+                     , cardano-sl
                      , cardano-sl-util
                      , cardano-sl-wallet-new
                      , data-default
@@ -649,57 +650,24 @@ test-suite integration
   type:               exitcode-stdio-1.0
   hs-source-dirs:     test
   main-is:            Integration.hs
-  other-modules:      Integration.Clients
-                      Integration.Fixtures
+  other-modules:      Integration.Fixtures
                       Integration.Expectations
                       Integration.Util
                       Integration.Specs.Addresses
 
-  default-extensions: BangPatterns
-                      ConstraintKinds
-                      DeriveGeneric
+  default-extensions: NoImplicitPrelude
                       FlexibleContexts
                       FlexibleInstances
-                      GADTs
                       InstanceSigs
                       LambdaCase
-                      MultiParamTypeClasses
-                      MultiWayIf
-                      NoImplicitPrelude
                       OverloadedStrings
                       RankNTypes
                       RecordWildCards
                       ScopedTypeVariables
-                      StandaloneDeriving
-                      TypeApplications
-                      TypeFamilies
 
   build-depends:      base
-                    , QuickCheck
-                    , bytestring
-                    , cardano-sl
-                    , cardano-sl-core
-                    , cardano-sl-crypto
-                    , cardano-sl-infra
-                    , cardano-sl-util
-                    , cardano-sl-wallet
+                    , cardano-sl-demo
                     , cardano-sl-wallet-new
-                    , containers
-                    , data-default
                     , hspec
-                    , http-client
-                    , http-types
-                    , lens
-                    , log-warper
-                    , mtl
                     , neat-interpolation
-                    , optparse-applicative
-                    , quickcheck-instances
-                    , random
-                    , stm
-                    , time
-                    , time-units
                     , universum
-                    , x509-store
-                    , parsec
-                    , attoparsec
